==> src/dspg17/ckelling/analysis/depsy/05_depsy_analysis.Rmd <==
---
title: "Analysis"
author: "Claire Kelling"
date: "6/27/2017"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

Now that I have all of the data on Depsy, I will perform analysis on both the nodal features and the edges. 

```{r load data}
load(file = "~/git/oss/data/oss/original/depsy/node_mat.Rdata")
load(file = "~/git/oss/data/oss/original/depsy/neighb_mat.Rdata")
```


First, I want to clean the data.
```{r}
neighb_mat <- as.data.frame(neighb_mat)
substrfunc <- function(x){substr(x, 6, nchar(as.character(x)))}
neighb_mat$neighbor <- sapply(neighb_mat$neighbor, substrfunc)

colnames(neighb_mat) <- c("Source","Target")
colnames(node_mat)[1] <- "Id"

save(neighb_mat, file = "~/git/oss/data/oss/working/depsy/neighb_mat.Rdata")
save(neighb_mat, file = "~/git/oss/src/ckelling/ScrapingCode/depsy/my_dat/neighb_mat.Rdata")

save(node_mat, file = "~/git/oss/data/oss/working/depsy/node_mat.Rdata")
save(node_mat, file = "~/git/oss/src/ckelling/ScrapingCode/depsy/my_dat/node_mat.Rdata")

```
